tasks/producer/requirements.txt
requests==2.32.3
pika==1.3.2
python-dotenv==1.1.0
aiohttp==3.11.18